[makerom] renamed dir.c/dir.h to romfs_fs.c/romfs_fs.h
